refactor for analysis and testability in ci
