+++ code/.coveragerc
@@ -1,0 +1,5 @@
+[run]
+relative_files = True
+include =
+  bluetooth_manager/*
+branch = True
